Refactor configurationutils and componentutils to inherit from a common base class
